rename: Hook 함수명과 파일명이 동일하도록 변경
