+++ src/components/Modal/CollectionModal/index.tsx
@@ -5,7 +5,7 @@ import '@/styles/scroll.css';
 import { useStore } from 'zustand';
 
 import Icon from '@/components/Icon/Icon';
-import useGetAllCollectionList from '@/hooks/serverStateHooks/useGetCollectionList';
+import useGetAllCollectionList from '@/hooks/serverStateHooks/useGetAllCollectionList';
 import modalStore from '@/stores/modalStore';
 
 import CollectionUnit from './CollectionUnit';
